+++ extensions/password-store/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Password Store Changelog
 
+## [Improvements] - 2026-02-24
+- Able to insert password via `pass insert`
+- Able to generate password via `pass generate`
+- Able to rename password via `pass mv`
+- Able to delete password via `pass rm`
+
 ## [Improvements] - 2025-03-17
 
 - Added ability to specify custom PATH to allow custom executable locations